backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
redis==5.0.1
chromadb==0.4.18
openai==1.3.7
tavily-python==0.3.1
apscheduler==3.10.4
websockets==12.0
python-dotenv==1.0.0
httpx==0.25.2
